re-writing pong, making it more manageable
