+++ frontend/src/components/auth/Login.tsx
@@ -29,6 +29,7 @@ const Login = () => {
 			navigate('/hub', {state: response.data})
 		} catch (error: any) {
 			console.error("Error:", error.response?.data || error.message);
+			navigate('/hub')
 			setLoginError("invalid username or password")
 			return;
 		}
